reworked BitString - bits array replaced with int (unsigned) deque. removed bitstring write/read cursors.

+++ cell/src/main/java/org/ton/java/cell/TopologicalOrderArray.java
@@ -1,10 +1,10 @@
 package org.ton.java.cell;
 
 public class TopologicalOrderArray {
-    byte[] cellHash;
+    int[] cellHash;
     Cell cell;
 
-    TopologicalOrderArray(byte[] cellHash, Cell cell) {
+    TopologicalOrderArray(int[] cellHash, Cell cell) {
         this.cellHash = cellHash;
         this.cell = cell;
     }
